basically reinstated the original home page from wokring authentification and then took the other home page and render it as a component upon login. new homepage component is named Aritist Search
