+++ client/src/pages/Home.tsx
@@ -1,88 +1,57 @@
-import { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
-import '../index.css';   
-import  Artist  from '../interfaces/artists';
+import { useState, useEffect, useLayoutEffect } from "react";
+import { retrieveUsers } from "../api/userAPI";
+import type { UserData } from "../interfaces/UserData";
+import ErrorPage from "./ErrorPage";
+import ArtistSearch from "../components/ArtistSearch";
+import auth from '../utils/auth';
 
-//const API_KEY =import.meta.env.LASTFM_API_KEY as string;
-
-
-
-const HomePage = () => {
-  const [artists, setArtists] = useState<Artist[]>([]);
+const Home = () => {
+  const [users, setUsers] = useState<UserData[]>([]);
+  const [error, setError] = useState(false);
+  const [loginCheck, setLoginCheck] = useState(false);
 
   useEffect(() => {
-    // Function to fetch artist info from Last.fm API
-    const fetchArtistImage = async (artistName: string) => {
-      try {
-        const response = await fetch(
-          `https://ws.audioscrobbler.com/2.0/?method=artist.getinfo&artist=${encodeURIComponent(artistName)}&api_key=045d5788747399c059939c53d984ca59&format=json`
-        );
-        const data = await response.json();
-        console.log('API response:', data); // Inspect the full API response
+    if (loginCheck) {
+      fetchUsers();
+    }
+  }, [loginCheck]);
 
-        if (data.artist && data.artist.image) {
-          // Find the image with the size 'extralarge'
-          const image = data.artist.image.find((img: any) => img.size === 'extralarge')?.['#text'];
-          return image || ''; // Return image or an empty string if not found
-        } else {
-          console.error('No image found for artist:', artistName);
-          return ''; // Return an empty string if no image is found
-        }
-      } catch (error) {
-        console.error('Error fetching artist image:', error);
-        return ''; // Return an empty string on error
-      }
-    };
-
-    // Fetch artist names and images
-    const loadArtistImages = async () => {
-      const artistNames = ['Ariana Grande', 'Taylor Swift', 'Ed Sheeran']; // List of artist names to display
-      const updatedArtists = await Promise.all(
-        artistNames.map(async (artistName) => {
-          const image = await fetchArtistImage(artistName);
-          return { name: artistName, image };
-        })
-      );
-      setArtists(updatedArtists);
-    };
-
-    loadArtistImages();
+  useLayoutEffect(() => {
+    checkLogin();
   }, []);
 
+  const checkLogin = () => {
+    if (auth.loggedIn()) {
+      setLoginCheck(true);
+    }
+  };
+
+  const fetchUsers = async () => {
+    try {
+      const data = await retrieveUsers();
+      setUsers(data);
+      console.log('Users:', data);
+    } catch (err) {
+      console.error('Failed to retrieve users:', err);
+      setError(true);
+    }
+  };
+
+  if (error) {
+    return <ErrorPage />;
+  }
+
   return (
-    <div className="container mt-5">
-      {/* Header */}
-      <header className="text-center">
-        <h1>Welcome to Discover-fy</h1>
-        <p>Your personal music playlist manager</p>
-      </header>
-
-      {/* Artist Images Section */}
-      <section className="artist-images text-center mt-4">
-        <h2>Discover New Artists</h2>
-        <div className="row mt-3">
-          {artists.map((artist, index) => (
-            <div key={index} className="col-md-3">
-              {artist.image ? (
-                <img src={artist.image} alt={artist.name} className="img-fluid" />
-              ) : (
-                <p>No image available</p>
-              )}
-              <p>{artist.name}</p>
-            </div>
-          ))}
+    <>
+      {!loginCheck ? (
+        <div className='login-notice'>
+          <h1>Login to view all your Playlists!</h1>
         </div>
-      </section>
-
-      {/* Call to Action Button */}
-      <section className="text-center mt-4">
-        <h3>Ready to discover your next playlist?</h3>
-        <Link to="/discover">
-          <button className="btn btn-primary btn-lg">Start Discovering</button>
-        </Link>
-      </section>
-    </div>
+      ) : (
+        <ArtistSearch />
+      )}
+    </>
   );
 };
 
-export default HomePage;
+export default Home;
